cluster/web-app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.1
blinker==1.8.2
Bootstrap-Flask==2.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
Flask-WTF==1.2.1
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kafka-python==2.0.2
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.0.0
pandas==2.2.2
protobuf==5.27.2
py4j==0.10.9.7
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.28.3
weaviate-client==4.6.5
Werkzeug==3.0.3
WTForms==3.1.2
